SLING-10421 validate configured and client supplied cookie domain value
add integration tests to cover the code changes
